works with transfer dialog. Now you can send tokens. Small string fix.

+++ app/src/main/java/com/blockchain/store/playmarket/ui/transfer_screen/transfer_info_screen/TransferInfoFragment.java
@@ -232,7 +232,7 @@ public class TransferInfoFragment extends Fragment implements TransferInfoContra
         if (tokenName != null) {
             String transferAmount = amountEditText.getText().toString();
             String totalBalance = balanceTextView.getText().toString();
-            if (Long.parseLong(transferAmount) < Long.parseLong(totalBalance)) {
+            if (Double.parseDouble(transferAmount) < Double.parseDouble(totalBalance)) {
                 return true;
             } else {
                 amountTextInputLayout.setError(getResources().getString(R.string.insufficient_funds));
